MISSION B/Waste-Classification-MLOps-Pipeline/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
prometheus-client==0.18.0
